requirements.txt
Django==1.7
django-appconf==1.0.1
django-classy-tags==0.6.2
django-easy-maps==0.9.2
django-registration-redux==1.3
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
djangorestframework==3.3.1
geopy==1.11.0
gunicorn==19.3.0
itsdangerous==0.24
requests==2.2.1
wheel==0.24.0
